Reduced the /sys tests to start again with hardcoded basic values.
To get the tests in a function state, it makes sense to get the tests
working first and then start DRY-ing up the code once all tests are
succeeding.

Also fixed the test for non-/sys environments and adapted it to the
current behaviour of this implementation, checking for the existence
of /sys/class/net.